Art and Hotfixes
I added the Animations and options for Bagel, Loaf and Baguette so they can now be played when selected from the player select screen. Merged some of the obstacles and buffs from Soo's project. Still need to add game manager and do adjustments to the camera and add Xingwei's end screen